[Dashboards] Register dashboard locator on server start (#218495)
## Summary

Registers the dashboard locator definition with the share plugin on the
Dashboard server start contract.

This moves Dashboard locator code from public to common so that we can
expose the Dashboard Locator definition on the server. The `getLocation`
method exposed by the locator does not appear to be used by any server
code. So the `getDashboardFilterFields` function provided to the
`DashboardLocatorDefinition` will throw a not supported error on the
server. This should currently never throw since getLocation is not
called by the server. But any future implementations where getLocation
is called on the server would throw an error and thus would need to
augment this function.

For posterity, here's an example where, in the future, we might need to
provide a scoped content management client to the DashboardLocatorParams
that can be used to load filters from a saved dashboard in the
`getDashboardFilterFields` function.

```ts
// in src/platform/plugins/shared/share/server/url_service/http/short_urls/register_create_route.ts

export const registerCreateRoute = (router: IRouter, url: ServerUrlService, contentManagement: ContentManagementServerSetup) => {

router.handleLegacyErrors(async (ctx, req, res) => {
  const savedObjects = (await ctx.core).savedObjects.client;

  const cmClient = contentManagement.contentClient
        .getForRequest({ request: req, requestHandlerContext: ctx })

  const shortUrls = url.shortUrls.get({ savedObjects });
  const { locatorId, params, slug } = req.body;
  const locator = url.locators.get(locatorId);
  // TODO find a way to pass cmClient to the locator. Maybe in params?

  try {
  const shortUrl = await shortUrls.create({
    locator,
    params,
    slug,
  });

  return res.ok({
    headers: {
      'content-type': 'application/json',
    },
    body: shortUrl.data,
  });
} catch (error) {
```

(cherry picked from commit 8e1a426c25497f332fd516dee46061c900618e96)

# Conflicts:
#	src/platform/plugins/shared/dashboard/public/dashboard_top_nav/dashboard_favorite_button.tsx
#	x-pack/platform/plugins/shared/ml/public/application/jobs/new_job/job_from_dashboard/quick_create_job_base.ts
