+++ src/platform/plugins/shared/dashboard/common/constants.ts
@@ -7,9 +7,10 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export enum PanelPlacementStrategy {
-  /** Place on the very top of the Dashboard, add the height of this panel to all other panels. */
-  placeAtTop = 'placeAtTop',
-  /** Look for the smallest y and x value where the default panel will fit. */
-  findTopLeftMostOpenSpace = 'findTopLeftMostOpenSpace',
-}
+export const LANDING_PAGE_PATH = '/list';
+export const DASHBOARD_APP_ID = 'dashboards';
+export const LEGACY_DASHBOARD_APP_ID = 'dashboard';
+export const SEARCH_SESSION_ID = 'searchSessionId';
+export const UI_SETTINGS = {
+  ENABLE_LABS_UI: 'labs:dashboard:enable_ui',
+};
