+++ src/platform/plugins/shared/dashboard/common/locator/locator.test.ts
