+++ src/platform/plugins/shared/dashboard/common/locator/locator.ts
@@ -11,13 +11,13 @@ import type { SerializableRecord } from '@kbn/utility-types';
 import { flow } from 'lodash';
 
 import type { Filter } from '@kbn/es-query';
-import { setStateToKbnUrl } from '@kbn/kibana-utils-plugin/public';
+import { setStateToKbnUrl } from '@kbn/kibana-utils-plugin/common';
 import type { LocatorDefinition, LocatorPublic } from '@kbn/share-plugin/public';
 import type { GlobalQueryStateFromUrl } from '@kbn/data-plugin/public';
 
 import { DASHBOARD_APP_LOCATOR } from '@kbn/deeplinks-analytics';
-import { DASHBOARD_APP_ID, SEARCH_SESSION_ID } from '../../plugin_constants';
-import { DashboardLocatorParams } from '../../dashboard_api/types';
+import { DashboardLocatorParams } from '../types';
+import { DASHBOARD_APP_ID, SEARCH_SESSION_ID } from '../constants';
 
 /**
  * Useful for ensuring that we don't pass any non-serializable values to history.push (for example, functions).
