+++ src/platform/plugins/shared/dashboard/common/types.ts
@@ -8,8 +8,21 @@
  */
 
 import type { Reference } from '@kbn/content-management-utils';
+import type { SerializableRecord } from '@kbn/utility-types';
+import type { Filter, Query, TimeRange } from '@kbn/es-query';
+import type { ViewMode } from '@kbn/presentation-publishing';
+import type { RefreshInterval } from '@kbn/data-plugin/public';
+import type {
+  ControlGroupRuntimeState,
+  ControlGroupSerializedState,
+} from '@kbn/controls-plugin/common';
+
 import type { DashboardPanelMap } from './dashboard_container/types';
-import type { DashboardAttributes } from '../server/content_management';
+import type {
+  DashboardAttributes,
+  DashboardOptions,
+  DashboardPanel,
+} from '../server/content_management';
 
 export interface DashboardCapabilities {
   showWriteControls: boolean;
@@ -32,3 +45,77 @@ export interface DashboardAttributesAndReferences {
   attributes: DashboardAttributes;
   references: Reference[];
 }
+
+export type DashboardSettings = DashboardOptions & {
+  description?: DashboardAttributes['description'];
+  tags: string[];
+  timeRestore: DashboardAttributes['timeRestore'];
+  title: DashboardAttributes['description'];
+};
+
+export interface DashboardState extends DashboardSettings {
+  query: Query;
+  filters: Filter[];
+  timeRange?: TimeRange;
+  refreshInterval?: RefreshInterval;
+  viewMode: ViewMode;
+  panels: DashboardPanelMap;
+
+  /**
+   * Temporary. Currently Dashboards are in charge of providing references to all of their children.
+   * Eventually this will be removed in favour of the Dashboard injecting references serverside.
+   */
+  references?: Reference[];
+
+  /**
+   * Serialized control group state.
+   * Contains state loaded from dashboard saved object
+   */
+  controlGroupInput?: ControlGroupSerializedState | undefined;
+  /**
+   * Runtime control group state.
+   * Contains state passed from dashboard locator
+   * Use runtime state when building input for portable dashboards
+   */
+  controlGroupState?: Partial<ControlGroupRuntimeState>;
+}
+
+export type DashboardLocatorParams = Partial<
+  Omit<DashboardState, 'panels' | 'controlGroupInput' | 'references'>
+> & {
+  /**
+   * If given, the dashboard saved object with this id will be loaded. If not given,
+   * a new, unsaved dashboard will be loaded up.
+   */
+  dashboardId?: string;
+
+  /**
+   * If not given, will use the uiSettings configuration for `storeInSessionStorage`. useHash determines
+   * whether to hash the data in the url to avoid url length issues.
+   */
+  useHash?: boolean;
+
+  /**
+   * When `true` filters from saved filters from destination dashboard as merged with applied filters
+   * When `false` applied filters take precedence and override saved filters
+   *
+   * true is default
+   */
+  preserveSavedFilters?: boolean;
+
+  /**
+   * Search search session ID to restore.
+   * (Background search)
+   */
+  searchSessionId?: string;
+
+  /**
+   * List of dashboard panels
+   */
+  panels?: Array<DashboardPanel & SerializableRecord>; // used SerializableRecord here to force the GridData type to be read as serializable
+
+  /**
+   * Control group changes
+   */
+  controlGroupState?: Partial<ControlGroupRuntimeState> & SerializableRecord; // used SerializableRecord here to force the GridData type to be read as serializable
+};
