+++ src/platform/plugins/shared/dashboard/public/dashboard_api/default_dashboard_state.ts
@@ -7,7 +7,7 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-import { DashboardState } from './types';
+import { DashboardState } from '../../common';
 
 export const DEFAULT_DASHBOARD_STATE: DashboardState = {
   viewMode: 'view',
