+++ src/platform/plugins/shared/dashboard/public/dashboard_api/load_dashboard_api.ts
@@ -8,9 +8,9 @@
  */
 
 import { ContentInsightsClient } from '@kbn/content-management-content-insights-public';
-import { DashboardPanelMap } from '../../common';
+import { DashboardPanelMap, DashboardState } from '../../common';
 import { getDashboardContentManagementService } from '../services/dashboard_content_management_service';
-import { DashboardCreationOptions, DashboardState, UnsavedPanelState } from './types';
+import { DashboardCreationOptions, UnsavedPanelState } from './types';
 import { getDashboardApi } from './get_dashboard_api';
 import { startQueryPerformanceTracking } from './performance/query_performance_tracking';
 import { coreServices } from '../services/kibana_services';
