+++ src/platform/plugins/shared/dashboard/public/dashboard_api/panels_manager.ts
@@ -29,14 +29,15 @@ import {
 } from '@kbn/presentation-publishing';
 import { i18n } from '@kbn/i18n';
 import { coreServices, usageCollectionService } from '../services/kibana_services';
-import { DashboardPanelMap, DashboardPanelState, prefixReferencesFromPanel } from '../../common';
+import { prefixReferencesFromPanel } from '../../common/dashboard_container/persistable_state/dashboard_container_references';
+import type { DashboardPanelMap, DashboardPanelState, DashboardState } from '../../common';
 import type { initializeTrackPanel } from './track_panel';
 import { getPanelAddedSuccessString } from '../dashboard_app/_dashboard_app_strings';
 import { runPanelPlacementStrategy } from '../panel_placement/place_new_panel_strategies';
 import { DEFAULT_PANEL_HEIGHT, DEFAULT_PANEL_WIDTH } from '../../common/content_management';
 import { DASHBOARD_UI_METRIC_ID } from '../utils/telemetry_constants';
 import { getDashboardPanelPlacementSetting } from '../panel_placement/panel_placement_registry';
-import { DashboardState, UnsavedPanelState } from './types';
+import { UnsavedPanelState } from './types';
 import { arePanelLayoutsEqual } from './are_panel_layouts_equal';
 import { dashboardClonePanelActionStrings } from '../dashboard_actions/_dashboard_actions_strings';
 import { placeClonePanel } from '../panel_placement/place_clone_panel_strategy';
