+++ src/platform/plugins/shared/dashboard/public/dashboard_api/save_modal/open_save_modal.tsx
@@ -21,7 +21,7 @@ import {
   savedObjectsTaggingService,
 } from '../../services/kibana_services';
 import { getDashboardContentManagementService } from '../../services/dashboard_content_management_service';
-import { DashboardState } from '../types';
+import { DashboardState } from '../../../common';
 import { DASHBOARD_CONTENT_ID, SAVED_OBJECT_POST_TIME } from '../../utils/telemetry_constants';
 import { extractTitleAndCount } from '../../utils/extract_title_and_count';
 import { DashboardSaveModal } from './save_modal';
