+++ src/platform/plugins/shared/dashboard/public/dashboard_api/view_mode_manager.ts
@@ -13,7 +13,7 @@ import { BehaviorSubject } from 'rxjs';
 import { LoadDashboardReturn } from '../services/dashboard_content_management_service/types';
 import { getDashboardBackupService } from '../services/dashboard_backup_service';
 import { getDashboardCapabilities } from '../utils/get_dashboard_capabilities';
-import { DashboardState } from './types';
+import { DashboardState } from '../../common';
 
 export function initializeViewModeManager(
   incomingEmbeddable?: EmbeddablePackageState,
