+++ src/platform/plugins/shared/dashboard/public/dashboard_app/dashboard_router.tsx
@@ -18,7 +18,7 @@ import React from 'react';
 import { render, unmountComponentAtNode } from 'react-dom';
 import { HashRouter, Redirect, RouteComponentProps } from 'react-router-dom';
 
-import { DASHBOARD_APP_ID, LANDING_PAGE_PATH } from '../plugin_constants';
+import { DASHBOARD_APP_ID, LANDING_PAGE_PATH } from '../../common/constants';
 import { RedirectToProps } from './types';
 import { coreServices, dataService, embeddableService } from '../services/kibana_services';
 import { getDashboardCapabilities } from '../utils/get_dashboard_capabilities';
