+++ src/platform/plugins/shared/dashboard/public/dashboard_app/listing_page/get_dashboard_list_item_link.ts
@@ -10,7 +10,7 @@
 import type { QueryState } from '@kbn/data-plugin/public';
 import { IKbnUrlStateStorage, setStateToKbnUrl } from '@kbn/kibana-utils-plugin/public';
 
-import { DASHBOARD_APP_ID } from '../../plugin_constants';
+import { DASHBOARD_APP_ID } from '../../../common/constants';
 import { coreServices } from '../../services/kibana_services';
 import { GLOBAL_STATE_STORAGE_KEY, createDashboardEditUrl } from '../../utils/urls';
