+++ src/platform/plugins/shared/dashboard/public/dashboard_app/top_nav/share/show_share_modal.test.tsx
@@ -8,12 +8,12 @@
  */
 
 import { Capabilities } from '@kbn/core/public';
-import { convertPanelMapToPanelsArray } from '../../../../common';
+import { convertPanelMapToPanelsArray } from '../../../../common/lib/dashboard_panel_converters';
 
 import { shareService } from '../../../services/kibana_services';
 import { showPublicUrlSwitch, ShowShareModal, ShowShareModalProps } from './show_share_modal';
 import { getDashboardBackupService } from '../../../services/dashboard_backup_service';
-import { DashboardLocatorParams, DashboardState } from '../../../dashboard_api/types';
+import { DashboardLocatorParams, DashboardState } from '../../../../common/types';
 
 describe('showPublicUrlSwitch', () => {
   test('returns false if "dashboard" app is not available', () => {
