+++ src/platform/plugins/shared/dashboard/public/dashboard_app/top_nav/share/show_share_modal.tsx
@@ -19,7 +19,8 @@ import { i18n } from '@kbn/i18n';
 import { getStateFromKbnUrl, setStateToKbnUrl, unhashUrl } from '@kbn/kibana-utils-plugin/public';
 
 import { FormattedMessage } from '@kbn/i18n-react';
-import { convertPanelMapToPanelsArray, DashboardPanelMap } from '../../../../common';
+import { convertPanelMapToPanelsArray } from '../../../../common/lib/dashboard_panel_converters';
+import { DashboardPanelMap } from '../../../../common';
 import {
   getDashboardBackupService,
   PANELS_CONTROL_GROUP_KEY,
@@ -28,7 +29,7 @@ import { coreServices, dataService, shareService } from '../../../services/kiban
 import { getDashboardCapabilities } from '../../../utils/get_dashboard_capabilities';
 import { shareModalStrings } from '../../_dashboard_app_strings';
 import { dashboardUrlParams } from '../../dashboard_router';
-import { DashboardLocatorParams } from '../../../dashboard_api/types';
+import { DashboardLocatorParams } from '../../../../common';
 
 const showFilterBarId = 'showFilterBar';
