+++ src/platform/plugins/shared/dashboard/public/dashboard_app/url/search_sessions_integration.ts
@@ -19,10 +19,11 @@ import { replaceUrlHashQuery } from '@kbn/kibana-utils-plugin/common';
 import type { Query } from '@kbn/es-query';
 import { SearchSessionInfoProvider } from '@kbn/data-plugin/public';
 import { DASHBOARD_APP_LOCATOR } from '@kbn/deeplinks-analytics';
-import { SEARCH_SESSION_ID } from '../../plugin_constants';
-import { convertPanelMapToPanelsArray } from '../../../common';
+import { SEARCH_SESSION_ID } from '../../../common/constants';
+import { convertPanelMapToPanelsArray } from '../../../common/lib/dashboard_panel_converters';
 import { dataService } from '../../services/kibana_services';
-import { DashboardApi, DashboardLocatorParams } from '../../dashboard_api/types';
+import { DashboardApi } from '../../dashboard_api/types';
+import { DashboardLocatorParams } from '../../../common/types';
 
 export const removeSearchSessionIdFromURL = (kbnUrlStateStorage: IKbnUrlStateStorage) => {
   kbnUrlStateStorage.kbnUrlControls.updateAsync((nextUrl) => {
