+++ src/platform/plugins/shared/dashboard/public/dashboard_listing/dashboard_listing.tsx
@@ -15,7 +15,7 @@ import { TableListViewKibanaProvider } from '@kbn/content-management-table-list-
 import { FormattedRelative, I18nProvider } from '@kbn/i18n-react';
 import { useExecutionContext } from '@kbn/kibana-react-plugin/public';
 
-import { DASHBOARD_APP_ID } from '../plugin_constants';
+import { DASHBOARD_APP_ID } from '../../common/constants';
 import { DASHBOARD_CONTENT_ID } from '../utils/telemetry_constants';
 import {
   coreServices,
