+++ src/platform/plugins/shared/dashboard/public/dashboard_top_nav/internal_dashboard_top_nav.tsx
@@ -27,7 +27,7 @@ import { TopNavMenuBadgeProps, TopNavMenuProps } from '@kbn/navigation-plugin/pu
 import { useBatchedPublishingSubjects } from '@kbn/presentation-publishing';
 import { LazyLabsFlyout, withSuspense } from '@kbn/presentation-util-plugin/public';
 
-import { UI_SETTINGS } from '../../common';
+import { UI_SETTINGS } from '../../common/constants';
 import { useDashboardApi } from '../dashboard_api/use_dashboard_api';
 import {
   dashboardManagedBadge,
@@ -39,7 +39,7 @@ import { useDashboardMountContext } from '../dashboard_app/hooks/dashboard_mount
 import { DashboardEditingToolbar } from '../dashboard_app/top_nav/dashboard_editing_toolbar';
 import { useDashboardMenuItems } from '../dashboard_app/top_nav/use_dashboard_menu_items';
 import { DashboardEmbedSettings, DashboardRedirect } from '../dashboard_app/types';
-import { LEGACY_DASHBOARD_APP_ID } from '../plugin_constants';
+import { LEGACY_DASHBOARD_APP_ID } from '../../common/constants';
 import { openSettingsFlyout } from '../dashboard_renderer/settings/open_settings_flyout';
 import { SaveDashboardReturn } from '../services/dashboard_content_management_service/types';
 import { getDashboardRecentlyAccessedService } from '../services/dashboard_recently_accessed_service';
