+++ src/platform/plugins/shared/dashboard/public/index.ts
@@ -10,17 +10,9 @@
 import { PluginInitializerContext } from '@kbn/core/public';
 
 import { DashboardPlugin } from './plugin';
-export {
-  DASHBOARD_APP_ID,
-  LEGACY_DASHBOARD_APP_ID,
-  PanelPlacementStrategy,
-} from './plugin_constants';
+export { PanelPlacementStrategy } from './plugin_constants';
 export { DASHBOARD_GRID_COLUMN_COUNT } from '../common/content_management';
-export type {
-  DashboardApi,
-  DashboardCreationOptions,
-  DashboardLocatorParams,
-} from './dashboard_api/types';
+export type { DashboardApi, DashboardCreationOptions } from './dashboard_api/types';
 export { DASHBOARD_API_TYPE } from './dashboard_api/types';
 export type { DashboardRendererProps } from './dashboard_renderer/dashboard_renderer';
 export { LazyDashboardRenderer as DashboardRenderer } from './dashboard_renderer/lazy_dashboard_renderer';
@@ -29,10 +21,12 @@ export type { DashboardStart } from './plugin';
 export { DashboardListingTable } from './dashboard_listing';
 export { DashboardTopNav } from './dashboard_top_nav';
 export type { RedirectToProps } from './dashboard_app/types';
-export { type DashboardAppLocator, cleanEmptyKeys } from './dashboard_app/locator/locator';
-export { getDashboardLocatorParamsFromEmbeddable } from './dashboard_app/locator/get_dashboard_locator_params';
 
 export { type SearchDashboardsResponse } from './services/dashboard_content_management_service/lib/find_dashboards';
+
+export { DASHBOARD_APP_ID } from '../common/constants';
+export { cleanEmptyKeys, DashboardAppLocatorDefinition } from '../common/locator/locator';
+export { getDashboardLocatorParamsFromEmbeddable } from '../common/locator/get_dashboard_locator_params';
 
 export function plugin(initializerContext: PluginInitializerContext) {
   return new DashboardPlugin(initializerContext);
