+++ src/platform/plugins/shared/dashboard/public/mocks.tsx
@@ -10,9 +10,9 @@
 import { ControlGroupApi } from '@kbn/controls-plugin/public';
 import { BehaviorSubject } from 'rxjs';
 import { DashboardStart } from './plugin';
-import { DashboardState } from './dashboard_api/types';
+import { DashboardState } from '../common/types';
 import { getDashboardApi } from './dashboard_api/get_dashboard_api';
-import { DashboardPanelState } from '../common';
+import { DashboardPanelState } from '../common/dashboard_container/types';
 
 export type Start = jest.Mocked<DashboardStart>;
