+++ src/platform/plugins/shared/dashboard/public/services/dashboard_backup_service.ts
@@ -16,7 +16,8 @@ import { set } from '@kbn/safer-lodash-set';
 
 import { SerializedPanelState, ViewMode } from '@kbn/presentation-publishing';
 import { coreServices, spacesService } from './kibana_services';
-import { DashboardState, UnsavedPanelState } from '../dashboard_api/types';
+import { DashboardState } from '../../common';
+import { UnsavedPanelState } from '../dashboard_api/types';
 import { DEFAULT_DASHBOARD_STATE } from '../dashboard_api/default_dashboard_state';
 
 export const DASHBOARD_PANELS_UNSAVED_ID = 'unsavedDashboard';
