+++ src/platform/plugins/shared/dashboard/server/ui_settings.ts
@@ -11,7 +11,7 @@ import { i18n } from '@kbn/i18n';
 import { schema } from '@kbn/config-schema';
 import { SETTING_CATEGORY } from '@kbn/presentation-util-plugin/server';
 import { UiSettingsParams } from '@kbn/core/types';
-import { UI_SETTINGS } from '../common';
+import { UI_SETTINGS } from '../common/constants';
 
 /**
  * uiSettings definitions for Dashboard.
