+++ x-pack/platform/plugins/shared/dashboard_enhanced/public/services/drilldowns/embeddable_to_dashboard_drilldown/embeddable_to_dashboard_drilldown.test.tsx
@@ -9,10 +9,10 @@ import { Filter, RangeFilter, FilterStateStore, Query, TimeRange } from '@kbn/es
 import { type Context, EmbeddableToDashboardDrilldown } from './embeddable_to_dashboard_drilldown';
 import { AbstractDashboardDrilldownConfig as Config } from '../abstract_dashboard_drilldown';
 import { savedObjectsServiceMock } from '@kbn/core/public/mocks';
-import { DashboardLocatorParams } from '@kbn/dashboard-plugin/public';
 import { StartDependencies } from '../../../plugin';
 import { StartServicesGetter } from '@kbn/kibana-utils-plugin/public/core';
-import { DashboardAppLocatorDefinition } from '@kbn/dashboard-plugin/public/dashboard_app/locator/locator';
+import type { DashboardLocatorParams } from '@kbn/dashboard-plugin/common';
+import { DashboardAppLocatorDefinition } from '@kbn/dashboard-plugin/public';
 import { BehaviorSubject } from 'rxjs';
 
 describe('.isConfigValid()', () => {
