+++ x-pack/solutions/security/plugins/security_solution/public/dashboards/components/dashboard_renderer.tsx
@@ -10,10 +10,10 @@ import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import type {
   DashboardApi,
   DashboardCreationOptions,
-  DashboardLocatorParams,
   DashboardRendererProps,
 } from '@kbn/dashboard-plugin/public';
 import { DashboardRenderer as DashboardContainerRenderer } from '@kbn/dashboard-plugin/public';
+import type { DashboardLocatorParams } from '@kbn/dashboard-plugin/common';
 import type { Filter, Query } from '@kbn/es-query';
 import type { ViewMode } from '@kbn/presentation-publishing';
